+++ packages/apollo-client/src/core/__tests__/QueryManager/index.ts
@@ -4,7 +4,11 @@ import { assign } from 'lodash';
 import gql from 'graphql-tag';
 import { DocumentNode, ExecutionResult, GraphQLError } from 'graphql';
 import { ApolloLink, Operation, Observable } from 'apollo-link';
-import { InMemoryCache, ApolloReducerConfig } from 'apollo-cache-inmemory';
+import {
+  InMemoryCache,
+  ApolloReducerConfig,
+  NormalizedCacheObject,
+} from 'apollo-cache-inmemory';
 
 // mocks
 import mockQueryManager from '../../../__mocks__/mockQueryManager';
@@ -118,7 +122,7 @@ describe('QueryManager', () => {
     });
     return new Promise<{
       result: ExecutionResult;
-      queryManager: QueryManager;
+      queryManager: QueryManager<NormalizedCacheObject>;
     }>((resolve, reject) => {
       queryManager
         .mutate({ mutation, variables })
@@ -2455,7 +2459,7 @@ describe('QueryManager', () => {
         },
       };
 
-      const queryManager = new QueryManager({
+      const queryManager = new QueryManager<NormalizedCacheObject>({
         link: mockSingleLink(
           {
             request: { query, variables },
@@ -3020,7 +3024,7 @@ describe('QueryManager', () => {
     });
 
     it('should only refetch once when we store reset', () => {
-      let queryManager: QueryManager;
+      let queryManager: QueryManager<NormalizedCacheObject>;
       const query = gql`
         query {
           author {
@@ -3080,7 +3084,7 @@ describe('QueryManager', () => {
     });
 
     it('should not refetch toredown queries', done => {
-      let queryManager: QueryManager;
+      let queryManager: QueryManager<NormalizedCacheObject>;
       let observable: ObservableQuery<any>;
       const query = gql`
         query {
@@ -3128,7 +3132,7 @@ describe('QueryManager', () => {
     });
 
     it('should not error on queries that are already in the store', () => {
-      let queryManager: QueryManager;
+      let queryManager: QueryManager<NormalizedCacheObject>;
       const query = gql`
         query {
           author {
@@ -3146,12 +3150,15 @@ describe('QueryManager', () => {
 
       let timesFired = 0;
       const link = ApolloLink.from([
-        () =>
-          new Observable(observer => {
-            timesFired += 1;
-            observer.next({ data });
-            return;
-          }),
+        new ApolloLink(
+          () =>
+            new Observable(observer => {
+              timesFired += 1;
+              observer.next({ data });
+              observer.complete();
+              return;
+            }),
+        ),
       ]);
       queryManager = createQueryManager({ link });
       const observable = queryManager.watchQuery<any>({
@@ -3163,13 +3170,27 @@ describe('QueryManager', () => {
       return observableToPromise(
         { observable, wait: 20 },
         result => {
-          expect(result.data).toEqual(data);
-          expect(timesFired).toBe(1);
-          setTimeout(queryManager.resetStore.bind(queryManager), 10);
+          try {
+            expect(result.data).toEqual(data);
+            expect(timesFired).toBe(1);
+          } catch (e) {
+            return fail(e);
+          }
+          setTimeout(async () => {
+            try {
+              await queryManager.resetStore();
+            } catch (e) {
+              fail(e);
+            }
+          }, 10);
         },
         result => {
-          expect(result.data).toEqual(data);
-          expect(timesFired).toBe(2);
+          try {
+            expect(result.data).toEqual(data);
+            expect(timesFired).toBe(2);
+          } catch (e) {
+            fail(e);
+          }
         },
       );
     });
@@ -3303,7 +3324,7 @@ describe('QueryManager', () => {
     });
 
     it('should throw an error on an inflight query() if the store is reset', done => {
-      let queryManager: QueryManager;
+      let queryManager: QueryManager<NormalizedCacheObject>;
       const query = gql`
         query {
           author {
@@ -3436,7 +3457,7 @@ describe('QueryManager', () => {
     });
 
     it('should only refetch once when we refetch observable queries', () => {
-      let queryManager: QueryManager;
+      let queryManager: QueryManager<NormalizedCacheObject>;
       const query = gql`
         query {
           author {
@@ -3496,7 +3517,7 @@ describe('QueryManager', () => {
     });
 
     it('should not refetch toredown queries', done => {
-      let queryManager: QueryManager;
+      let queryManager: QueryManager<NormalizedCacheObject>;
       let observable: ObservableQuery<any>;
       const query = gql`
         query {
@@ -3544,7 +3565,7 @@ describe('QueryManager', () => {
     });
 
     it('should not error on queries that are already in the store', () => {
-      let queryManager: QueryManager;
+      let queryManager: QueryManager<NormalizedCacheObject>;
       const query = gql`
         query {
           author {
@@ -3756,7 +3777,7 @@ describe('QueryManager', () => {
     });
 
     it('should NOT throw an error on an inflight query() if the observed queries are refetched', done => {
-      let queryManager: QueryManager;
+      let queryManager: QueryManager<NormalizedCacheObject>;
       const query = gql`
         query {
           author {
@@ -4368,7 +4389,7 @@ describe('QueryManager', () => {
       );
       const cache = new InMemoryCache();
 
-      const queryManager = new QueryManager({
+      const queryManager = new QueryManager<NormalizedCacheObject>({
         link,
         store: new DataStore(cache),
       });