+++ packages/apollo-cache-inmemory/src/__tests__/cache.ts
@@ -847,19 +847,7 @@ describe('Cache', () => {
         `,
       });
 
-      expect((proxy as InMemoryCache).extract()).toEqual({
-        foo: {
-          e: 4,
-          h: {
-            type: 'id',
-            id: 'bar',
-            generated: false,
-          },
-        },
-        bar: {
-          i: 7,
-        },
-      });
+      expect((proxy as InMemoryCache).extract()).toMatchSnapshot();
       proxy.writeFragment({
         data: { __typename: 'Foo', f: 5, g: 6, h: { id: 'bar', j: 8, k: 9 } },
         id: 'foo',
@@ -875,23 +863,7 @@ describe('Cache', () => {
         `,
       });
 
-      expect((proxy as InMemoryCache).extract()).toEqual({
-        foo: {
-          e: 4,
-          f: 5,
-          g: 6,
-          h: {
-            type: 'id',
-            id: 'bar',
-            generated: false,
-          },
-        },
-        bar: {
-          i: 7,
-          j: 8,
-          k: 9,
-        },
-      });
+      expect((proxy as InMemoryCache).extract()).toMatchSnapshot();
 
       proxy.writeFragment({
         data: { i: 10, __typename: 'Bar' },
@@ -903,23 +875,7 @@ describe('Cache', () => {
         `,
       });
 
-      expect((proxy as InMemoryCache).extract()).toEqual({
-        foo: {
-          e: 4,
-          f: 5,
-          g: 6,
-          h: {
-            type: 'id',
-            id: 'bar',
-            generated: false,
-          },
-        },
-        bar: {
-          i: 10,
-          j: 8,
-          k: 9,
-        },
-      });
+      expect((proxy as InMemoryCache).extract()).toMatchSnapshot();
 
       proxy.writeFragment({
         data: { j: 11, k: 12, __typename: 'Bar' },
@@ -932,23 +888,7 @@ describe('Cache', () => {
         `,
       });
 
-      expect((proxy as InMemoryCache).extract()).toEqual({
-        foo: {
-          e: 4,
-          f: 5,
-          g: 6,
-          h: {
-            type: 'id',
-            id: 'bar',
-            generated: false,
-          },
-        },
-        bar: {
-          i: 10,
-          j: 11,
-          k: 12,
-        },
-      });
+      expect((proxy as InMemoryCache).extract()).toMatchSnapshot();
 
       proxy.writeFragment({
         data: {
@@ -980,23 +920,7 @@ describe('Cache', () => {
         fragmentName: 'fooFragment',
       });
 
-      expect((proxy as InMemoryCache).extract()).toEqual({
-        foo: {
-          e: 4,
-          f: 5,
-          g: 6,
-          h: {
-            type: 'id',
-            id: 'bar',
-            generated: false,
-          },
-        },
-        bar: {
-          i: 7,
-          j: 8,
-          k: 9,
-        },
-      });
+      expect((proxy as InMemoryCache).extract()).toMatchSnapshot();
 
       proxy.writeFragment({
         data: { __typename: 'Bar', i: 10, j: 11, k: 12 },
@@ -1022,23 +946,7 @@ describe('Cache', () => {
         fragmentName: 'barFragment',
       });
 
-      expect((proxy as InMemoryCache).extract()).toEqual({
-        foo: {
-          e: 4,
-          f: 5,
-          g: 6,
-          h: {
-            type: 'id',
-            id: 'bar',
-            generated: false,
-          },
-        },
-        bar: {
-          i: 10,
-          j: 11,
-          k: 12,
-        },
-      });
+      expect((proxy as InMemoryCache).extract()).toMatchSnapshot();
     });
     it('writes data that can be read back', () => {
       const proxy = createCache({