+++ packages/apollo-client/CHANGELOG.md
@@ -2,8 +2,11 @@
 # Change log
 
 ### vNEXT
+- Added the graphQLResultHasError in QueryManager.ts to check not only if result.errors is null, but also empty.
 - Errors occurring during fetchMore no longer errors out the original query [PR#2906](https://github.com/apollographql/apollo-client/pull/2906)
 - Map coverage to original source
+
+### 2.2.4
 - Added `getCacheKey` function to the link context for use in state-link [PR#2998](https://github.com/apollographql/apollo-client/pull/2998)
 - Fix Memory Leak in Query Manager [PR#3119](https://github.com/apollographql/apollo-client/pull/3119)
 - Pass non-optimistic query to `subscribeToMore` updateQuery